+++ example/lib/example_colors.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:light_html_editor/light_html_editor.dart';
 
-class RegularExample extends StatelessWidget {
-  const RegularExample({Key? key}) : super(key: key);
+class ColorExample extends StatelessWidget {
+  const ColorExample({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -21,7 +21,9 @@ class RegularExample extends StatelessWidget {
             onChanged: (String html) {
               // do something with the richtext
             },
-            alwaysShowButtons: true,
+            availableColors: [
+              "#affe11", // only one color available to the editor
+            ],
           ),
         ),
       ),