+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:example/example_external_controller.dart';
 import 'package:example/example_regular.dart';
 import 'package:flutter/material.dart';