+++ lib/light_html_editor.dart
@@ -2,3 +2,5 @@ export 'editor.dart';
 export 'placeholder.dart';
 export 'renderer.dart';
 export 'api/parser.dart';
+export 'data/editor_properties.dart';
+export 'data/renderer_properties.dart';